renamed connect to connect_request
